refactor: extract useHostServer, useHostPlayers from Host.vue
